+++ pakyow-assets/CHANGELOG.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+  * [fixed] CDN prefix is now correctly added to assets in plugin views
+
 # 1.0
 
   * Hello, Web